Move ForkTsCheckerWebpackPlugin from @nuxt/webpack `base` to `client`
